requirements.txt
AWSIoTPythonSDK
numpy
pysocks